+++ webapp/frontend/lib/settings/settings_service.dart
@@ -2,6 +2,9 @@ import 'dart:convert';
 
 import 'package:flutter/material.dart';
 import 'package:get_10101/common/http_client.dart';
+import 'package:json_annotation/json_annotation.dart';
+
+part 'settings_service.g.dart';
 
 class SettingsService {
   const SettingsService();
@@ -20,9 +23,20 @@ class SettingsService {
     final response = await HttpClientManager.instance.get(Uri(path: '/api/seed'));
 
     if (response.statusCode == 200) {
-      return jsonDecode(response.body);
+      var jsonData = jsonDecode(response.body);
+      return Seed.fromJson(jsonData).seed;
     } else {
       throw FlutterError("Failed to fetch seed phrase");
     }
   }
+}
+
+@JsonSerializable()
+class Seed {
+  final List<String> seed;
+
+  Seed({required this.seed});
+
+  factory Seed.fromJson(Map<String, dynamic> json) => _$SeedFromJson(json);
+  Map<String, dynamic> toJson() => _$SeedToJson(this);
 }