+++ webapp/frontend/lib/settings/settings_service.g.dart
@@ -1,0 +1,15 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+part of 'settings_service.dart';
+
+// **************************************************************************
+// JsonSerializableGenerator
+// **************************************************************************
+
+Seed _$SeedFromJson(Map<String, dynamic> json) => Seed(
+      seed: (json['seed'] as List<dynamic>).map((e) => e as String).toList(),
+    );
+
+Map<String, dynamic> _$SeedToJson(Seed instance) => <String, dynamic>{
+      'seed': instance.seed,
+    };